Remove Gemfile.lock
This code is not an application we want to be able to reproducibly
deploy, but a module added to an existing environment without it's
dependencies being managed.

So we do not need a Gemfile.lock because it is only used in CI and the
requirements in the Gemfile should be enough.

This should prevent robots from asking us to update every dependency
when a new release is done.